Change the hub's position information in thrusterStateEffector
The hub's position and velocity information depends on what gravitational body is set as zeroBase. However, the hub's "hubPosition" and "hubVelocity" states are always written with respect to the SPICE origin, and do not account for the zeroBase command. This can lead to errors of orbital magnitudes, particularly when attaching the thruster to a body other than the hub.